Move arithmetic to its own file.
